Use composite object for master list and master list line domain
