+++ domain/src/lib.rs
@@ -1,5 +1,3 @@
-pub mod master_list;
-pub mod master_list_line;
 pub mod shipment_tax_update;
 
 use chrono::{NaiveDate, NaiveDateTime};
