+++ repository/src/db_diesel/master_list_line.rs
@@ -6,13 +6,20 @@ use crate::{
         MasterListLineRow,
     },
 };
-use domain::{master_list_line::MasterListLineFilter, Pagination};
+use domain::{EqualFilter, Pagination};
 
 use super::{DBType, StorageConnection};
 
 use diesel::prelude::*;
 
 pub type MasterListLine = MasterListLineRow;
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct MasterListLineFilter {
+    pub id: Option<EqualFilter<String>>,
+    pub item_id: Option<EqualFilter<String>>,
+    pub master_list_id: Option<EqualFilter<String>>,
+}
 
 pub struct MasterListLineRepository<'a> {
     connection: &'a StorageConnection,
@@ -74,4 +81,29 @@ fn create_filtered_query(
     }
 
     Ok(query)
+}
+
+impl MasterListLineFilter {
+    pub fn new() -> MasterListLineFilter {
+        MasterListLineFilter {
+            id: None,
+            item_id: None,
+            master_list_id: None,
+        }
+    }
+
+    pub fn id(mut self, filter: EqualFilter<String>) -> Self {
+        self.id = Some(filter);
+        self
+    }
+
+    pub fn item_id(mut self, filter: EqualFilter<String>) -> Self {
+        self.item_id = Some(filter);
+        self
+    }
+
+    pub fn master_list_id(mut self, filter: EqualFilter<String>) -> Self {
+        self.master_list_id = Some(filter);
+        self
+    }
 }
