+++ repository/src/db_diesel/mod.rs
@@ -48,8 +48,8 @@ pub use item_stats::*;
 pub use key_value_store::*;
 pub use location::*;
 pub use location_row::LocationRowRepository;
-pub use master_list::{MasterList, MasterListRepository};
-pub use master_list_line::{MasterListLine, MasterListLineRepository};
+pub use master_list::*;
+pub use master_list_line::*;
 pub use master_list_line_row::MasterListLineRowRepository;
 pub use master_list_name_join::MasterListNameJoinRepository;
 pub use master_list_row::MasterListRowRepository;
