+++ repository/src/schema/mod.rs
@@ -52,8 +52,8 @@ pub use item_is_visible::ItemIsVisibleRow;
 pub use item_stats::*;
 pub use key_value_store::*;
 pub use location::LocationRow;
-pub use master_list::MasterListRow;
-pub use master_list_line::MasterListLineRow;
+pub use master_list::*;
+pub use master_list_line::*;
 pub use master_list_name_join::MasterListNameJoinRow;
 pub use name::NameRow;
 pub use name_store_join::NameStoreJoinRow;
