+++ repository/src/tests.rs
@@ -340,20 +340,17 @@ mod repository_test {
         },
         test_db, CentralSyncBufferRepository, ChangelogRepository, InvoiceLineRepository,
         InvoiceLineRowRepository, InvoiceRepository, ItemRepository, ItemStatsFilter,
-        ItemStatsRepository, KeyValueStoreRepository, MasterListLineRepository,
+        ItemStatsRepository, KeyValueStoreRepository, MasterListFilter, MasterListLineRepository,
         MasterListLineRowRepository, MasterListNameJoinRepository, MasterListRepository,
         MasterListRowRepository, NameFilter, NameQueryRepository, NameRepository, NameSort,
         NameSortField, NumberRowRepository, OutboundShipmentRepository, RequisitionFilter,
         RequisitionLineFilter, RequisitionLineRepository, RequisitionLineRowRepository,
         RequisitionRepository, RequisitionRowRepository, StockLineFilter, StockLineRepository,
-        StockLineRowRepository, StocktakeRowRepository, StoreRowRepository, UserAccountRepository,
+        StockLineRowRepository, StocktakeRowRepository, StoreRowRepository, UserAccountRepository, MasterListLineFilter,
     };
     use chrono::{Duration, Utc};
     use diesel::{sql_query, sql_types::Text, RunQueryDsl};
-    use domain::{
-        master_list::MasterListFilter, master_list_line::MasterListLineFilter, DateFilter,
-        EqualFilter, Pagination, SimpleStringFilter,
-    };
+    use domain::{DateFilter, EqualFilter, Pagination, SimpleStringFilter};
     use util::constants::NUMBER_OF_DAYS_IN_A_MONTH;
 
     #[actix_rt::test]
