+++ server/tests/graphql/master_lists.rs
@@ -1,12 +1,9 @@
 mod graphql {
     use crate::graphql::assert_graphql_query;
-    use domain::{
-        master_list::{MasterListFilter, MasterListSort},
-        EqualFilter, PaginationOption, SimpleStringFilter,
-    };
+    use domain::{EqualFilter, PaginationOption, SimpleStringFilter};
     use repository::{
         mock::{mock_master_list_master_list_line_filter_test, MockDataInserts},
-        MasterList, StorageConnectionManager,
+        MasterList, StorageConnectionManager, MasterListFilter, MasterListSort,
     };
     use serde_json::{json, Value};
     use server::test_utils::setup_all;
